fix(image): move opaqueimageview to new
