+++ src/app/Components/ArtworkGrids/ArtworkGridItem.tsx
@@ -6,6 +6,7 @@ import { ArtworksFiltersStore } from "app/Components/ArtworkFilter/ArtworkFilter
 import { DurationProvider } from "app/Components/Countdown"
 import OpaqueImageView from "app/Components/OpaqueImageView/OpaqueImageView"
 
+import { OpaqueImageView as NewOpaqueImageView } from "app/Components/OpaqueImageView2"
 import { GlobalStore, useFeatureFlag } from "app/store/GlobalStore"
 import { PageableRouteProps } from "app/system/navigation/useNavigateToPageableRoute"
 import { useArtworkBidding } from "app/utils/Websockets/auctions/useArtworkBidding"
@@ -16,7 +17,7 @@ import {
   PlaceholderRaggedText,
   RandomNumberGenerator,
 } from "app/utils/placeholders"
-import { OpaqueImageView as NewOpaqueImageView, Touchable } from "palette"
+import { Touchable } from "palette"
 import React, { useRef } from "react"
 import { View } from "react-native"
 import { createFragmentContainer, graphql } from "react-relay"
