+++ src/app/Components/Home/ArtistRails/ArtistCard.tsx
@@ -1,9 +1,10 @@
 import { CloseIcon, Flex, Text, useColor } from "@artsy/palette-mobile"
 import { ArtistCard_artist$data } from "__generated__/ArtistCard_artist.graphql"
+import { OpaqueImageView } from "app/Components/OpaqueImageView2"
 
 import { navigate } from "app/system/navigation/navigate"
 import { extractNodes } from "app/utils/extractNodes"
-import { FollowButton, OpaqueImageView as NewOpaqueImageView, Touchable } from "palette"
+import { FollowButton, Touchable } from "palette"
 import { createFragmentContainer, graphql } from "react-relay"
 import styled from "styled-components/native"
 
@@ -141,7 +142,7 @@ const ArtworkCardImages = ({
                 alignItems="center"
                 overflow="hidden"
               >
-                <NewOpaqueImageView
+                <OpaqueImageView
                   imageURL={image.src}
                   height={containerHeight * ZOOM_IN_FACTOR}
                   width={imageWidth * ZOOM_IN_FACTOR}
