+++ src/app/Components/OpaqueImageView2/OpaqueImageView2.tsx
