+++ src/app/Components/OpaqueImageView2/index.ts
