+++ src/app/Scenes/Artwork/Components/ImageCarousel/ImageWithLoadingState.tsx
@@ -1,4 +1,4 @@
-import OpaqueImageView2 from "palette/elements/OpaqueImageView/OpaqueImageView2"
+import { OpaqueImageView } from "app/Components/OpaqueImageView2"
 import React from "react"
 import { TouchableWithoutFeedback, View, ViewProps } from "react-native"
 
@@ -30,7 +30,7 @@ export const ImageWithLoadingState = React.forwardRef<View, ImageWithLoadingStat
     return (
       <TouchableWithoutFeedback onPress={onPress} accessibilityLabel="Image with Loading State">
         <View style={[{ width, height }, props.style]} ref={ref}>
-          <OpaqueImageView2
+          <OpaqueImageView
             useRawURL
             onLoad={() => {
               if (props.onLoad) {
