+++ src/app/Scenes/MyCollection/Screens/Insights/MedianSalePriceAtAuction.tsx
@@ -1,6 +1,7 @@
 import { OwnerType } from "@artsy/cohesion"
 import { Spacer, NoArtworkIcon, Flex, Text } from "@artsy/palette-mobile"
 import { MedianSalePriceAtAuctionQuery } from "__generated__/MedianSalePriceAtAuctionQuery.graphql"
+import { OpaqueImageView } from "app/Components/OpaqueImageView2"
 import {
   PlaceholderBox,
   PlaceholderText,
@@ -9,7 +10,7 @@ import {
 } from "app/utils/placeholders"
 import { ProvideScreenTrackingWithCohesionSchema } from "app/utils/track"
 import { screen } from "app/utils/track/helpers"
-import { OpaqueImageView, Touchable } from "palette"
+import { Touchable } from "palette"
 import { Suspense, useCallback, useEffect, useRef, useState } from "react"
 import { ScrollView } from "react-native"
 import { graphql, useLazyLoadQuery } from "react-relay"
