+++ src/app/Scenes/Search/components/TrendingArtistCard.tsx
@@ -1,6 +1,6 @@
 import { Flex, Text } from "@artsy/palette-mobile"
 import { TrendingArtistCard_artist$key } from "__generated__/TrendingArtistCard_artist.graphql"
-import { OpaqueImageView } from "palette"
+import { OpaqueImageView } from "app/Components/OpaqueImageView2"
 import { TouchableHighlight } from "react-native"
 import { useFragment } from "react-relay"
 import { graphql } from "relay-runtime"
