+++ src/app/Scenes/ViewingRoom/ViewingRoomArtworks.tsx
@@ -2,6 +2,7 @@ import { Flex, Box, useSpace, Text } from "@artsy/palette-mobile"
 import { ViewingRoomArtworksQueryRendererQuery } from "__generated__/ViewingRoomArtworksQueryRendererQuery.graphql"
 import { ViewingRoomArtworks_viewingRoom$data } from "__generated__/ViewingRoomArtworks_viewingRoom.graphql"
 import ImageView from "app/Components/OpaqueImageView/OpaqueImageView"
+import { OpaqueImageView } from "app/Components/OpaqueImageView2"
 import { ReadMore } from "app/Components/ReadMore"
 import { useFeatureFlag } from "app/store/GlobalStore"
 import { navigate } from "app/system/navigation/navigate"
@@ -9,7 +10,7 @@ import { defaultEnvironment } from "app/system/relay/createEnvironment"
 import { extractNodes } from "app/utils/extractNodes"
 import renderWithLoadProgress from "app/utils/renderWithLoadProgress"
 import { ProvideScreenTracking, Schema } from "app/utils/track"
-import { OpaqueImageView, Separator, Spinner, Touchable } from "palette"
+import { Separator, Spinner, Touchable } from "palette"
 import React, { useMemo, useState } from "react"
 import { FlatList, useWindowDimensions } from "react-native"
 import { createPaginationContainer, graphql, QueryRenderer, RelayPaginationProp } from "react-relay"
