+++ src/palette/elements/index.tsx
@@ -14,7 +14,6 @@ export * from "./Join"
 export * from "./LineGraph"
 export * from "./List"
 export * from "./Message"
-export * from "./OpaqueImageView"
 export * from "./Pill"
 export * from "./PopIn"
 export * from "./ProgressBar"
